+++ index.html
@@ -54,6 +54,7 @@
  <footer>
   <h2>Copyright</h2>
  </footer>
- <script src="./script.js"></script>
+ <script src="https://cdnjs.cloudflare.com/ajax/libs/luxon/2.0.2/luxon.min.js"></script>
+ <script type="module" src="./script.js"></script>
 </body>
 </html>
